AI can win against Medium difficulty. Still a long way to go

+++ src/com/sillysoft/lux/agent/utils/simulation/PathSimulationResult.java
@@ -1,11 +1,11 @@
 package com.sillysoft.lux.agent.utils.simulation;
 
 public class PathSimulationResult {
-    public final int countriesCaptured;
-    public final int armiesLeft;
+    public final double countriesCaptured;
+    public final double armiesLeft;
     public final boolean reachedContinent;
 
-    public PathSimulationResult(int countriesCaptured, int armiesLeft, boolean reachedContinent) {
+    public PathSimulationResult(double countriesCaptured, double armiesLeft, boolean reachedContinent) {
         this.countriesCaptured = countriesCaptured;
         this.armiesLeft = armiesLeft;
         this.reachedContinent = reachedContinent;
